+++ src_files/src/src_server/server_comandos/cmd_enemigo_despliega_unidad.h
@@ -8,11 +8,11 @@ private:
     uint8_t id_jugador;
     uint8_t id_unidad;
     uint8_t tipo_unidad;
-    long tiempo_entrenamiento;
+    uint16_t tiempo_entrenamiento;
     Coordenadas& coords_spawn;
 
 public:
-    CmdEnemigoDespliegaUnidadServer(uint8_t id_unidad, uint8_t id_jugador, uint8_t tipo_unidad, long tiempo_entrenamiento, Coordenadas& coords_spawn);
+    CmdEnemigoDespliegaUnidadServer(uint8_t id_unidad, uint8_t id_jugador, uint8_t tipo_unidad, uint16_t tiempo_entrenamiento, Coordenadas& coords_spawn);
     virtual void enviarComando(ProtocoloServidor& protocolo) const override;
     virtual ~CmdEnemigoDespliegaUnidadServer() = default;
 };
